d) Added Seach and Filter option, Added more constraints for creating instances | e) Added GUI for Search and Filter option
d) Added Seach and Filter option
Added a textfield to search the tableView.
Added a comboBox to select the type of attribute to search through.

d) Added more constraints for creating instances
Added a BooleanBinding to grey out the Save Button when not all attributes are set while creating a new instance. This prevents setting attributes to null.

e) Added GUI for Search and Filter option
Added a textfield to input the search string and a comboBox to select the attribute to search through